feat(adapter-agents-settings): adds adapter and request funcion to agents settings update
