+++ src/api/agentsSettings/adapters.ts
@@ -1,0 +1,9 @@
+
+export interface AgentsSettingsUpdateResponse {
+  message: string;
+  error: string;
+};
+
+export function adapterAgentsSettingsUpdate(response: AgentsSettingsUpdateResponse) {
+  return response;
+}
