+++ src/services/features.service.ts
@@ -3,6 +3,8 @@ import { VTEXFetch } from "../utils/VTEXFetch";
 import storeProvider from "../store/provider.store";
 import { setDisableFeatureLoading, setFeatureList, setFeatureLoading, setIntegratedFeatures, setUpdateFeatureLoading } from "../store/projectSlice";
 import { agentsList, integratedAgentsList } from "../api/agents/requests";
+import { agentsSettingsUpdate } from "../api/agentsSettings/requests";
+import { SettingsFormData } from "../components/settings/SettingsContainer/SettingsContext";
 
 export async function updateFeatureList() {
   const availableFeatures = await agentsList();
@@ -48,20 +50,11 @@ export async function integrateFeature(feature_uuid: string, project_uuid: strin
   }
 }
 
-export async function updateAgentSettings(body: any) {
+export async function updateAgentSettings({ agentUuid, code, formData }: { agentUuid: string, code: 'abandoned_cart' | 'order_status', formData: SettingsFormData }) {
   storeProvider.dispatch(setFeatureLoading(true))
 
   try {
-    const response = await VTEXFetch<{
-      message: string;
-      error: string;
-    }>(`/_v/update-feature-settings`, {
-      method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(body),
-    });
+    const response = await agentsSettingsUpdate({ agentUuid, code, formData });
 
     if (!response || response.error) {
       throw new Error(response?.message || 'error updating agent.');
